% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map the locations of accidents from FARS data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numeric indicating a state number.}

\item{year}{A year to plot; it should be a numeric.}
}
\value{
This function makes a plot that represents locations of accidents.
If there are no matches of \code{state.num} in a data set, it results in an error.
}
\description{
This function maps the locations of accidents using latitude and longitude information
from the FARS data files.
}
\details{
\code{\link[maps]{map}} is utilized to draw a state map.
}
\examples{
\dontrun{fars_map_state(1, 2015)}

}
