% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read files into a list of data frame tables}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numeric vector of years that you are interested in.}
}
\value{
This function returns a list of data frame tables. If there are any
inappropriate inputs, it leads to a warning message and returns \code{NULL}.
}
\description{
Read files into a list of data frame tables
}
\examples{
\dontrun{fars_read_years(c(2013, 2015))}

}
