% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This function summarizes a list of data frames. You can check the number of
cases per year every month.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numeric vector of years that you are interested in.}
}
\value{
This function returns a data frame table(\code{tbl_df}) summarizing data.
One of the column names is "MONTH", and the others are years coerced to characters.
If any element of the input vector does not match a file list, it gives you a warning message.
}
\description{
This function summarizes a list of data frames. You can check the number of
cases per year every month.
}
\examples{
\dontrun{fars_summarize_years(2014:2015)}

}
