% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A year used to  make a file name. This should be a numeric.}
}
\value{
Returns a plot of the accidents based on the \code{state.num} and
\code{year} inputs. Returns an error if the state or year do not exist in the
data set.
}
\description{
This function creates an appropriate file name for a given year.
}
\examples{
\dontrun{make_filename(2013)} #returns "accident_2013.csv.bz2"

}
